+++ DVSTracker/interval.h
@@ -4,11 +4,11 @@
 class Interval
 {
 public:
-    Interval(unsigned int ts = 0, int posX = -1, int posY = -1, unsigned int dT = 0);
-    unsigned int timeStamp;
+    Interval(unsigned int ts = 0, int posX = -1, int posY = -1, double dT = 0);
+    double timeStamp;
     int x;
     int y;
-    unsigned int deltaT;
+    double deltaT;
     float weight;
 };
